refactor: separate features to folders
