+++ helpers/index.js
@@ -1,0 +1,3 @@
+export { cli_options, findIndex } from './cliTool.js';
+export { CustomError, errorHandler } from './errorHandler.js';
+export const { exit, stdin, stdout, stderr, argv } = process;
