+++ mainStreams/index.js
@@ -1,0 +1,3 @@
+export { input_stream } from './inputStream.js';
+export { output_stream } from './outputStream.js';
+export { getCiphersStreams } from './transformStream.js';
